test: put all unit tests in sources
